Parser moved to HCIdump thread
